gh-113317, AC: Move warn() and fail() to libclinic.errors
