+++ Tools/clinic/libclinic/__init__.py
@@ -2,6 +2,8 @@ from typing import Final
 
 from .errors import (
     ClinicError,
+    warn,
+    fail,
 )
 from .formatting import (
     SIG_END_MARKER,
@@ -32,6 +34,8 @@ from .utils import (
 __all__ = [
     # Error handling
     "ClinicError",
+    "warn",
+    "fail",
 
     # Formatting helpers
     "SIG_END_MARKER",
